+++ ios/TwilioVoiceReactNative+CallKit.m
@@ -167,6 +167,7 @@ NSString * const kCustomParametersKeyDisplayName = @"displayName";
     TVOConnectOptions *connectOptions = [TVOConnectOptions optionsWithAccessToken:self.accessToken block:^(TVOConnectOptionsBuilder *builder) {
         builder.params = self.twimlParams;
         builder.uuid = uuid;
+        builder.callMessageDelegate = self;
     }];
     TVOCall *call = [TwilioVoiceSDK connectWithOptions:connectOptions delegate:self];
     if (call) {
@@ -183,6 +184,7 @@ NSString * const kCustomParametersKeyDisplayName = @"displayName";
     TVOCallInvite *callInvite = self.callInviteMap[uuid.UUIDString];
     TVOAcceptOptions *acceptOptions = [TVOAcceptOptions optionsWithCallInvite:callInvite block:^(TVOAcceptOptionsBuilder *builder) {
         builder.uuid = uuid;
+        builder.callMessageDelegate = self;
     }];
 
     TVOCall *call = [callInvite acceptWithOptions:acceptOptions delegate:self];